+++ changelog.md
@@ -17,6 +17,7 @@ __[changes]__
 __[maintenance]__
 
 * Update certificate information for Blender LTS installer.
+* Update certificate information for CMake installers.
 * Update certificate information for Node.js installer.
 * NLog library is updated from 6.0.1 to 6.0.2.